requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-google
pandas
pyspark==3.5.5
apache-airflow-providers-apache-spark
